+++ CHANGELOG.md
@@ -1,4 +1,16 @@
 # Changelog
+## v2.3.0
+
+#### Breaking changes
+- Changed the exit code from 1 to 0 for `ls` when used with an empty bucket. Exits with 1 if the bucket is non-existent. ([#722](https://github.com/peak/s5cmd/issues/722))
+
+#### Features
+- Added prefix and wildcard support to `cat` command. ([#716](https://github.com/peak/s5cmd/issues/716))
+- Added `head` command. ([#730](https://github.com/peak/s5cmd/pull/730))
+
+#### Bugfixes
+- Fixed the `cp` command to work with the `--content-type` flag when performing a copy operation from S3 to S3. ([#738](https://github.com/peak/s5cmd/issues/738))
+
 ## v2.2.2 - 13 Sep 2023 
 
 #### Bugfixes